Test: 당첨 번호, 보너스 번호, 추첨 번호 테스트 고도화
